backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.9.0  # Supabase Python client (includes postgrest and httpx dependencies)

# Data validation and serialization
pydantic>=2.10.0
pydantic-settings>=2.6.0

# LLM and AI
google-generativeai==0.3.2  # Google Gemini (FREE - default)
anthropic==0.18.1           # Anthropic Claude (optional)
openai==1.12.0              # OpenAI (optional)
groq==0.4.1                 # Groq (optional)

# Environment and configuration
python-dotenv==1.0.1

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-rerunfailures==14.0
hypothesis==6.98.3

# Code quality
autopep8==2.0.4
flake8==7.0.0

# Utilities
python-dateutil==2.8.2
numpy==1.26.3
